gdbus-test-codegen: Allocate full integer space for uint16 properties
Because there is no G_TYPE_[U]INT16, we need to use integer types,
otherwise we'll end up overwriting memory (in this case on the stack).

+++ gio/tests/gdbus-test-codegen.c
@@ -703,8 +703,8 @@ check_bar_proxy (FooBar    *proxy,
   ClientData *data;
   guchar val_y;
   gboolean val_b;
-  gint16 val_n;
-  guint16 val_q;
+  gint val_n;
+  guint val_q;
   gint val_i;
   guint val_u;
   gint64 val_x;
